+++ src/json.rs
@@ -1,17 +1,17 @@
-use redis::{RedisResult, Value};
+use redis::{ParsingError, Value};
 use serde::de::DeserializeOwned;
 
 /// Json struct is a wrapper to handle the return types from the RedisJSON commands.
-/// 
+///
 /// RedisJSON usually returns values in square brackets, which you usually had to handle manually:
-/// 
+///
 /// ```rust,no_run
 /// # use redis::{Client, JsonCommands, RedisResult};
 /// # use redis_macros::{FromRedisValue, ToRedisArgs, Json};
 /// # use serde::{Deserialize, Serialize};
 /// # #[derive(Serialize, Deserialize)]
 /// # struct User { id: u32 }
-/// # 
+/// #
 /// # fn main () -> redis::RedisResult<()> {
 /// # let client = redis::Client::open("redis://localhost:6379/")?;
 /// # let mut con = client.get_connection()?;
@@ -21,10 +21,10 @@ use serde::de::DeserializeOwned;
 /// # Ok(())
 /// # }
 /// ```
-/// 
+///
 /// Instead, `Json` implements the `FromRedisValue` trait, removes the square brackets and deserializes from JSON.
 /// For this your type don't even have to implement `FromRedisValue`, it only requires to be serde `Deserialize`-able.
-/// 
+///
 /// ```rust,no_run
 /// # use redis::{Client, JsonCommands, RedisResult};
 /// # use redis_macros::Json;
@@ -40,8 +40,8 @@ use serde::de::DeserializeOwned;
 /// let Json(user): Json<User> = con.json_get("user", "$")?;  // => User { id: 1 }
 /// # Ok(())
 /// # }
-/// ``` 
-/// 
+/// ```
+///
 /// This command is designed to use RedisJSON commands. You could probably use this type
 /// to parse normal command outputs, but it removes the first and last character
 /// so it is not recommended.
@@ -49,15 +49,15 @@ use serde::de::DeserializeOwned;
 #[derive(Debug)]
 pub struct Json<T>(
     /// The inner type to deserialize
-    pub T
+    pub T,
 );
 
 impl<T> ::redis::FromRedisValue for Json<T>
 where
     T: DeserializeOwned,
 {
-    fn from_redis_value(v: &Value) -> RedisResult<Self> {
-        match *v {
+    fn from_redis_value(v: Value) -> Result<Json<T>, ParsingError> {
+        match v {
             Value::BulkString(ref bytes) => {
                 if let Ok(s) = ::std::str::from_utf8(bytes) {
                     let mut ch = s.chars();
@@ -65,32 +65,18 @@ where
                         if let Ok(t) = serde_json::from_str(ch.as_str()) {
                             Ok(Json(t))
                         } else {
-                            Err(::redis::RedisError::from((
-                                ::redis::ErrorKind::TypeError,
-                                "Response was of incompatible type",
-                                format!("Response type in JSON was not deserializable. (response was {v:?})"),
-                            )))
+                            Err(format!("Response type in JSON was not deserializable. (response was {v:?})").into())
                         }
                     } else {
-                        Err(::redis::RedisError::from((
-                            ::redis::ErrorKind::TypeError,
-                            "Response was of incompatible type",
-                            format!("Response type was not JSON type. (response was {v:?})"),
-                        )))
+                        Err(format!("Response type was not JSON type. (response was {v:?})").into())
                     }
                 } else {
-                    Err(::redis::RedisError::from((
-                        ::redis::ErrorKind::TypeError,
-                        "Response was of incompatible type",
-                        format!("Response was not valid UTF-8 string. (response was {v:?})"),
-                    )))
+                    Err(format!("Response was not valid UTF-8 string. (response was {v:?})").into())
                 }
             }
-            _ => Err(::redis::RedisError::from((
-                ::redis::ErrorKind::TypeError,
-                "Response was of incompatible type",
-                format!("Response type not RedisJSON deserializable. (response was {v:?})"),
-            ))),
+            _ => Err(
+                format!("Response type not RedisJSON deserializable. (response was {v:?})").into(),
+            ),
         }
     }
 }