+++ Changelog.md
@@ -2,9 +2,16 @@
 
 ## [Unreleased]
 
+## [1.0.0-rc.0]
+
 ### Updated
 
-- Update dependencies
+- **Breaking change**: Update Redis version to 1.0.0-rc.0
+    - This library mostly can be used without any changes, but read the [redis-rs breaking changes](https://github.com/redis-rs/redis-rs/blob/redis-1.0.0-rc.0/version1.md)
+    - Implement `ToSingleRedisArg` for types deriving the `ToSingleRedis` macro (see [ToSingleRedisArg trait](https://github.com/redis-rs/redis-rs/blob/redis-1.0.0-rc.0/version1.md#tosingleredisarg-trait-breaking-change))
+    - Change `FromRedisValue` to accept owned types (see [More efficient deserialization](https://github.com/redis-rs/redis-rs/blob/redis-1.0.0-rc.0/version1.md#more-efficient-deserialization-breaking-change))
+    - Change `FromRedisValue` error from `RedisError` to `ParsingError` ([Better error messages](https://github.com/redis-rs/redis-rs/blob/redis-1.0.0-rc.0/version1.md#better-error-messages-potentially-breaking))
+- Update other dependencies
 
 ## [0.5.6] - 2025-07-20